modul/datadasar/requirements.txt
fastapi
pydantic
uvicorn
load_dotenv
python-multipart
mysql-connector-python